Move enums and ui_states to new data packages
