+++ app/src/main/java/com/dubproductions/bracket/data/status/MatchStatus.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket.data
+package com.dubproductions.bracket.data.status
 
 enum class MatchStatus(val status: String) {
     PENDING("pending"),
