+++ app/src/main/java/com/dubproductions/bracket/data/status/TournamentStatus.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket.data
+package com.dubproductions.bracket.data.status
 
 enum class TournamentStatus(val status: String) {
     REGISTERING("registering"),
