+++ app/src/main/java/com/dubproductions/bracket/data/ui_state/ParticipantsUIState.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket.ui_state
+package com.dubproductions.bracket.data.ui_state
 
 import com.dubproductions.bracket.data.Participant
