+++ app/src/main/java/com/dubproductions/bracket/data/ui_state/TournamentCreationUIState.kt
@@ -1,4 +1,4 @@
-package com.dubproductions.bracket.ui_state
+package com.dubproductions.bracket.data.ui_state
 
 data class TournamentCreationUIState(
     var successfulCreation: Boolean? = null,
