+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/TournamentCreation.kt
@@ -29,7 +29,7 @@ import com.dubproductions.bracket.R
 import com.dubproductions.bracket.Type
 import com.dubproductions.bracket.Validation
 import com.dubproductions.bracket.ui.components.ReusableDialog
-import com.dubproductions.bracket.ui_state.TournamentCreationUIState
+import com.dubproductions.bracket.data.ui_state.TournamentCreationUIState
 
 @Composable
 fun TournamentCreationScreen(
