+++ app/src/main/java/com/dubproductions/bracket/ui/main/hosting/participant/ParticipantMatches.kt
@@ -13,7 +13,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.dubproductions.bracket.data.Match
-import com.dubproductions.bracket.data.MatchStatus
+import com.dubproductions.bracket.data.status.MatchStatus
 import com.dubproductions.bracket.data.Participant
 import com.dubproductions.bracket.ui.components.DeclareWinnerDialog
 import com.dubproductions.bracket.ui.components.MatchCard
