+++ app/src/main/java/com/dubproductions/bracket/viewmodel/CreationViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import com.dubproductions.bracket.data.Participant
 import com.dubproductions.bracket.data.Tournament
 import com.dubproductions.bracket.data.repository.TournamentRepositoryImpl
-import com.dubproductions.bracket.ui_state.TournamentCreationUIState
+import com.dubproductions.bracket.data.ui_state.TournamentCreationUIState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
