+++ app/src/main/java/com/dubproductions/bracket/viewmodel/RegistrationViewModel.kt
@@ -3,7 +3,7 @@ package com.dubproductions.bracket.viewmodel
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import com.dubproductions.bracket.data.repository.TournamentRepositoryImpl
-import com.dubproductions.bracket.ui_state.RegistrationUIState
+import com.dubproductions.bracket.data.ui_state.RegistrationUIState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
